♻️refactor: Intensity 관련 공통 컴포넌트 분리 및 적용
