+++ src/app/user/[id]/setting/edit/components/EditIntensity/EditIntensity.tsx
@@ -1,11 +1,12 @@
-import { Check } from "@/components/icons";
+"use client";
+
 import * as GS from "../../UserEdit.styles";
 import * as S from "./EditIntensity.styles";
-import useTheme from "@/lib/hooks/useTheme";
-import { UseFormRegister, UseFormRegisterReturn } from "react-hook-form";
+import { UseFormRegister } from "react-hook-form";
 import { UserEditData } from "../../UserEdit.types";
 import { INTENSITY_OPTIONS } from "@/lib/constants/variable";
 import { IntensityOption } from "@/types/OriginDataType";
+import { UserIntensityItem } from "@/components";
 
 interface EditIntensityProps {
   register: UseFormRegister<UserEditData>;
@@ -22,7 +23,7 @@ const EditIntensity = ({ register, selectedIntensity }: EditIntensityProps) => {
 
       <S.IntensityOptionList>
         {INTENSITY_OPTIONS.map(({ label, value }) => (
-          <IntensityItem
+          <UserIntensityItem
             key={value}
             value={value}
             optionDescription={label}
@@ -36,37 +37,3 @@ const EditIntensity = ({ register, selectedIntensity }: EditIntensityProps) => {
 };
 
 export default EditIntensity;
-
-interface IntensityItemProps {
-  key?: string | number;
-  value: string;
-  optionDescription: string;
-  isSelected: boolean;
-  register: UseFormRegisterReturn;
-}
-
-const IntensityItem = ({ isSelected, value, optionDescription, register }: IntensityItemProps) => {
-  const theme = useTheme();
-
-  return (
-    <S.IntensityItemContainer>
-      <input
-        id={`intensity_${value}`}
-        type="radio"
-        style={{ display: "none" }}
-        value={value}
-        {...register}
-      />
-      <S.IntensityItemLabel htmlFor={`intensity_${value}`}>
-        <S.IntensityItemCircle $isSelected={isSelected}>
-          <Check
-            className={`w-6 h-6 mx-auto my-auto transition-colors`}
-            stroke={isSelected ? theme?.white_100 : theme?.gray_300}
-            strokeWidth={3.5}
-          />
-        </S.IntensityItemCircle>
-        <S.IntensityOptionDescription>{optionDescription}</S.IntensityOptionDescription>
-      </S.IntensityItemLabel>
-    </S.IntensityItemContainer>
-  );
-};
