+++ src/components/index.ts
@@ -15,3 +15,4 @@ export { default as Window } from "./Window";
 export { default as Textarea } from "./Textarea";
 export { default as InputRadio } from "./InputRadio/InputRadio";
 export { default as UserEditInput } from "./UserEditInput/UserEditInput";
+export { default as UserIntensityItem } from "./UserIntensityItem/UserIntensityItem";
